Make CrateContext::new() fn.
Move construction of a CrateContext into a static method on
CrateContext.
